perf: remove field teamId
