+++ packages/global/support/wallet/sub/type.d.ts
@@ -52,25 +52,6 @@ export type TeamSubSchema = {
   currentExtraDatasetSize: number;
 };
 
-export type TeamCouponSub = {
-  type: SubTypeEnum; // 套餐类型
-  startTime: string; // 开始时间
-  expiredTime: string; // 结束时间
-  price: number; // 价格
-  level: StandardSubLevelEnum; // 套餐等级
-  extraDatasetSize: number; // 额外知识库容量
-  totalPoints: number; // 总积分
-  surplusPoints: number; // 剩余积分
-};
-
-export type TeamCouponSchema = {
-  teamId: string;
-  key: string;
-  subscriptions: TeamCouponSub[];
-  redeemedAt?: Date;
-  expiredAt?: Date;
-};
-
 export type FeTeamPlanStatusType = {
   [SubTypeEnum.standard]?: TeamSubSchema;
   standardConstants?: TeamStandardSubPlanItemType;
