+++ projects/app/src/pages/account/info/index.tsx
@@ -9,10 +9,9 @@ import {
   Link,
   Progress,
   Grid,
-  BoxProps,
-  VStack,
-  Text
+  BoxProps
 } from '@chakra-ui/react';
+import { RedeemCouponModal } from './RedeemCouponModal';
 import { useForm } from 'react-hook-form';
 import { UserUpdateParams } from '@/types/user';
 import { useToast } from '@fastgpt/web/hooks/useToast';
@@ -624,53 +623,6 @@ const PlanUsage = () => {
       />
     </Box>
   ) : null;
-};
-
-const RedeemCouponModal = ({
-  isOpen,
-  onClose,
-  onSuccess
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-  onSuccess: () => void;
-}) => {
-  const { t } = useTranslation();
-
-  const [couponCode, setCouponCode] = React.useState('');
-
-  const { runAsync: redeemCouponAsync, loading } = useRequest2(redeemCoupon, {
-    manual: true,
-    onSuccess: () => {
-      onSuccess();
-      onClose();
-    },
-    successToast: t('common.Success')
-  });
-
-  return (
-    <MyModal
-      isOpen={isOpen}
-      onClose={onClose}
-      iconSrc="support/account/coupon"
-      title={t('account_info:redeem_coupon')}
-    >
-      <VStack px={'24px'} py={'36px'}>
-        <Text fontWeight={500} fontSize={'14px'} mb={'8px'} mr={'auto'}>
-          {t('account_info:redeem_coupon')}
-        </Text>
-        <Input
-          mb={'24px'}
-          placeholder={t('account_info:redeem_coupon')}
-          value={couponCode}
-          onChange={(e) => setCouponCode(e.target.value)}
-        />
-        <Button isLoading={loading} ml={'auto'} onClick={() => redeemCouponAsync(couponCode)}>
-          {t('account_info:confirm')}
-        </Button>
-      </VStack>
-    </MyModal>
-  );
 };
 
 const ButtonStyles = {
